mode responsive

+++ delete-article.php
@@ -1,11 +1,20 @@
 <?php
 require_once __DIR__ . '/data/data-base.php';
-$articleDB = require __DIR__ . '/./data/models/ArticleDB.php';
+require_once __DIR__ . '/data/security.php';
+$currentUser = isLoggedin();
 
-$_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
-$id = $_GET['id'] ?? '';
-if ($id) {
-    $articleDB->deleteOne($id);
+if ($currentUser) {
+
+    $articleDB = require __DIR__ . '/./data/models/ArticleDB.php';
+
+    $_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
+    $id = $_GET['id'] ?? '';
+    if ($id) {
+        $article = $articleDB->fetch($id);
+        if ($article['author'] === $currentUser['id']) {
+            $articleDB->deleteOne($id);
+        }
+    }
 }
 
 header('Location: /');
